rolled back edit to hdf5 generator

+++ biondi/dataset.py
@@ -1555,7 +1555,7 @@ def tile_sample_hdf5_generator_v2(wsi_filename, im_size=1024, sample_size=100, n
         f_old = h5py.File(filename + '1.hdf5', 'r')
         p = f_old['full_randomized_tile_indices']
     tile_stack = []
-    for index in p[:int(sample_size * (max_previous + 1))]:
+    for index in p[:sample_size * (max_previous + 1)]:
         # determine row in WSI
         i = index // grid_width
         # determine column in WSI
